+++ CHANGELOG.rst
@@ -4,9 +4,10 @@ Changelog
 0.2.0
 -----
 
-- Add `llms_txt_max_lines` configuration option to limit `llms-full.txt` file size
+- Add `llms_txt_full_max_size` configuration option to limit `llms-full.txt` file size
 - Automatically add content from `.. include::` directives
 - Add path resolution for certain directives
+- Add `llms.txt` file option, with `llms_txt_title` and `llms_txt_summary` config values
 
 0.1.0
 -----