+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Changelog
 -----
 
 - Add `llms_txt_full_max_size` configuration option to limit `llms-full.txt` file size
-- Automatically add content from `.. include::` directives
-- Add path resolution for certain directives
+- Automatically add content from `include` directives in  `llms-full.txt`
+- Add path resolution for a given set of directives  in `llms-full.txt`
 - Add `llms.txt` file option, with `llms_txt_title` and `llms_txt_summary` config values
 
 0.1.0